+++ popcornn/paths/__init__.py
@@ -1,6 +1,5 @@
 from .mlp import MLPpath
 from .linear import LinearPath
-from .base_path import Images
 
 path_dict = {
     "mlp" : MLPpath,
