+++ popcornn/tools/__init__.py
@@ -4,4 +4,4 @@ from .integrator import ODEintegrator
 from .logging import logging
 from .ase import pair_displacement, output_to_atoms, wrap_points
 from .metrics import Metrics
-from .molecules import process_images
+from .images import Images, process_images
